// DataCorner/TD/src/app/interfaces/view-form-interface.ts
export interface ComponentProperties {
    batchOptions: string[];
    categoryOptions: string[];
    locationOptions: string[];
    skillSetOptions: string[];
    educationOptions: string[];
    projectidOptions: string[];
  }
 export interface RoleInfo {
    roleId: number;
    roleName: string;
    defaultColumns: string;
  }